refactor(navigation): extract Routes to a separate file and update imports
