+++ shared/presentation/ui/src/commonMain/kotlin/com/almarpa/kmmtemplateapp/presentation/ui/navigation/NavigationActions.kt
@@ -4,27 +4,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import com.almarpa.kmmtemplateapp.domain.models.Pokemon
-import kotlinx.serialization.Serializable
-
-/**
- * Destinations used throughout the app.
- */
-sealed interface Routes {
-    @Serializable
-    data object Splash : Routes
-
-    @Serializable
-    data object PokemonList : Routes
-
-    @Serializable
-    data object Team : Routes
-
-    @Serializable
-    data object Settings : Routes
-
-    @Serializable
-    data class Detail(val pokemon: Pokemon) : Routes
-}
+import com.almarpa.kmmtemplateapp.presentation.ui.navigation.routes.Routes
 
 /**
  * Models the navigation actions in the app.
